[FLINK-14441][Table-Common]Fix ValueLiteralExpression#getValueAs when valueClass is Period.class and given class is Integer

+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/expressions/ValueLiteralExpression.java
@@ -104,9 +104,9 @@ public final class ValueLiteralExpression implements ResolvedExpression {
 			convertedValue = Duration.ofMillis(longVal);
 		}
 
-		else if (valueClass == Period.class && clazz == Long.class) {
+		else if (valueClass == Period.class && clazz == Integer.class) {
 			final Period period = (Period) value;
-			convertedValue = period.toTotalMonths();
+			convertedValue = (int)period.toTotalMonths();
 		}
 
 		else if (valueClass == Integer.class && clazz == Period.class) {
